Separated server internal files
