+++ auth/login.php
@@ -2,7 +2,7 @@
 const _EXEC = 1;
 
 require_once($_SERVER['DOCUMENT_ROOT'] . '/vendor/autoload.php');
-require_once($_SERVER['DOCUMENT_ROOT'] . '/server/skautis.secret.php');
+require_once($_SERVER['DOCUMENT_ROOT'] . '/server/internal/skautis.secret.php');
 
 $skautis = Skautis\Skautis::getInstance(SKAUTIS_APP_ID, SKAUTIS_TEST_MODE);
 $prefix = 'https://odymaterialy.skauting.cz';
