+++ server/internal/.htaccess
@@ -1,0 +1,4 @@
+Order Deny,Allow
+Deny from all
+Allow from 127.0.0.1
+
